+++ Level5-Module0/src/_01_Lambda_Methods/LambdaMethods.java
