--- Level5-Module1/src/_01_Olympic_Rings/OlympicRings_Threaded.java
@@ -1,8 +1,0 @@
-package _01_Olympic_Rings;
-
-
-public class OlympicRings_Threaded {
-	// Make A Program that uses Threads and robots to draw the Olympic rings. One robot should draw one ring simultaneously with the other 4 robots.
-
-}
-
